requirements.txt
textblob
python-docx
PyPDF2
nltk
textstat
